[lang] accept filter keywords that are not quoted (#35)
* [lang] accept filter keywords that are not quoted

To make usage a little more friendly, I'd like to add support
for filters that are not quoted and make the search case-
insensitive.  I tried to follow the SumoLogic syntax that
allows for some punctuation.  Wildcards are also allowed in
unquoted filters now.

Files:
  * Cargo.toml: Bump regex version to something more recent.
  * README.md: Filters no longer need to be quoted.
  * lang.rs: Add a keyword parser that does not require quotes.
    Add a Keyword struct used to contain the search string and
    the regex that can be used to do the matching/capturing.
  * lib.rs, operator.rs: Use the Keyword struct from lang.

* Multiple filters are now supported

* Remove MatchAll/MatchFilter in favor of returning a list of keywords
so that we can support multiple keywords.  Currently, the keywords
are expected to be ANDed together.  Future work would be to support
'OR'/'NOT'.  If a filter keyword is '*' or an empty string, it will
be removed from the list.  In other words, the old MatchAll behavior
is now represented as an empty list.

The Keyword struct has also been reworked to not hold onto the Regex.
Instead, the upper layers can now convert the Keyword to a Regex and
drop the Keyword object altogether.

* Convert all filter Keywords to Regexes and use those to check for
matches on inputs.  If the list of keywords is empty, then everything
matches.

* Pass the Regex to the Parse constructor instead of a str and store it
directly in the Parse struct.

* Test that an extra '*' does not affect anything and that multiple
filter keywords work.