[operators] add limit operator (#40)
* [operators] add limit operator

One of the first operators I reached for after trying to use this
tool was 'limit'.  Fortunately, it wasn't implemented yet, so I
got to play around a bit (!).  This change contains the first part
of the limit implementation, which passes records through up to
the limit and then drops the remaining records.  In a followup
change, I'll address aborting the upstream after the limit has
been reached so we don't have to read until the end of the input
stream.

I had to make some non-trivial changes to the current design
to allow for limit to work:

  * Allow non-aggregate operators to maintain state since limit
    needs to keep a counter in "head" mode.
  * Since non-agg operators now maintain state, the PreAggAdapter
    needs to re-instantiate the encapsulated operator to reset
    the state.  For example, if 'limit 5' is placed after an
    aggregate operator, we only want to show two lines of output
    from the aggregate.  To do this, an OperatorBuilder trait has
    been added that can be used to construct the mutable operator
    state from the operator definition.
  * The existing operators have been changed to implement a new
    UnaryPreAggFunction trait since they do not maintain state and
    it's easier to create a generic OperatorBuilder for these types
    of traits.
  * The error checks done during pre-agg operators needs to be
    done separately since we don't want them to fail construction
    in the PreAggAdapter.

I haven't fixed the issue with a limit operator after an aggregate
since I don't have a feel for what the best design might be.

If you think I am taking things in the wrong direction, let me know
and I can rework it.

Files:
  * lang.rs: Add "limit" to the grammar.
  * lib.rs: Move the conversion of the types from the lang
    module to the typecheck module.  The idea being that
    lang builds the raw syntax tree and typecheck does the
    semantic analysis phase.  So, any errors should be
    caught in there.
  * operator.rs: Moved TypeError out to the typecheck mod.
    Add the Limit operator and definition (LimitDef).
    Changed the operator's process() method to take self
    as mutable so they can maintain state.  The operator
    constructors no longer return Result since the checks
    should now be done in the typecheck module.
  * typecheck.rs: Do more type checking in this module.

* Incorporate review feedback