[lang] first cut at improving parse error messages (#44)
* [lang] first cut at improving parse error messages

Informative parser error messages can go a long way towards making
this tool easier to use.  This change tries to lay some groundwork
for catching common errors and reporting meaningful error messages.

First, the nom_locate crate has been added and nom's "verbose-errors"
feature has been enabled so that more information can be gained
about were parse errors occur and where values are coming from in
the query.  I've also started changing the grammar to make it more
lax so that more problems are treated as semantic errors instead
of parse errors.  For example, having the parse fail because there
are no arguments passed to count_distinct is not friendly.  Instead,
the parse now succeeds and the error is raised later on.  In the
future, I'd like to raise that type of error even later so that
a set of fields to use can be suggested.  Reporting errors is done
through the annotate-snippets crate which generates nice error
messages like in rustc.

The only errors I'm tackling in this first pass are unterminated
quotes and parentheses, as well as the count_distinct() aggregate
function.

Files:
  * Cargo.toml: Enable the verbose-errors feature in nom so that
    we can get more error info.  The nom_locate crate makes it a
    little easier to track the position of values in the query.
    The annotate-snippets library makes it easy to report errors
    in a pretty manner.
  * agrind.rs: Print parse errors to stderr.
  * lang.rs: Use nom_locate's LocatedSpan as the input type.
    Start to rework the grammar to be more lax.  Added some
    infrastructure for reporting errors.
  * lib.rs, typecheck.rs: Move the building of aggregate functions
    into typecheck to match how inline operators work.

* Incorporating feedback

* missed a 'cargo fmt'

* some more fixups