[panic] handle stdout being closed gracefully
Redirecting the output of agrind to "head" results in a
panic since the result of the writeln!() is unwrapped.
This change checks the result and drops out of the render
thread.  The main thread then bombs out when trying to
send records to the renderer.

Not sure the best way to test this...

Files:
  * lib.rs: Check for errors from the renderer and drop
    out of the loopers.
  * render.rs: Return a Result from render()