[data] add support for dates and function calls (#112)
* [data] add support for dates and function calls

This change adds the boilerplate for handling date-time
values.  The chrono crate is used for dealing with values
since that seems to be the popular one at the moment.
Since the only way to get a date into the pipeline is by
using the parseDate() function, I've updated the expression
grammar to support function calls.  To help flesh out the
function call infrastructure, I've added a bunch of math
and a couple of string functions as well.

Future Work:
  - Make all expressions Positioned<> for better errors
  - Add more functions
  - Support user-supplied date formats in parseDate()

Files:
  * Cargo.toml: Add a couple of date-related packages
  * data.rs: Add Value::DateTime variant
  * funcs.rs: The implementations for expression functions
  * lang.rs: Add a parser for function calls
  * operator.rs: Handle function calls
  * typecheck.rs: Convert function calls and their args
    from lang to operator types

Related to #10

* [data] add some more tests and refactor static function eval

* fix copy-pasta for expected arg count and add tests