[lang] migrate to nom 6 (#148)
* [lang] migrate to nom 6

!THIS IS NOT COMPLETE! I'm mostly looking for some high-level
feedback at the moment.  Most of the tests are passing, except
for the alias-related ones.

This change migrates the lang module from nom 4 to nom 6 (the
latest release at the time).  I tried to follow the existing
code for the most part.  I didn't want to deviate too much
so that we could still compare the two impls and not introduce
too many new bugs.  The changes I have introduced are to try
and improve error handling.  I tried to follow the advice in
this post:

  https://eyalkalderon.com/blog/nom-error-recovery/

Basically, the parse should always succeed and errors are
inserted into the AST (e.g. a bad expression is represented
as an Expr::Error in the tree).  Error messages are logged
using the existing ErrorReporter mechanism.  In addition,
when unexpected data is encountered, the parser tries to
seek forward to the next section of the pipeline (i.e. the
vertical bar).

TODO:
 * Implement aliases
 * Try to handle more error cases
 * Remove dead code
 * Double check that old parser behaviors are preserved

* [tests] disable alias tests

* [lang] rename expect_bar() to expect_pipe()

* [lang] support for aliases

* [fmt] missed a fmt run

* [lang] address some comments and fix some clippy issues

* [lang] remove lifetime from QueryContainer