Maintain ordering of columns when emitting aggregation JSON.
Before this patch, the JsonPrinter for data:Aggregate asks serde_json
to serialize its `Vec<VMap>` data field.  Because VMap is currently
defined to be a HashMap, the ordering of keys in the map will
effectively be random on a per-row basis.  Changing VMap to use an
ordered map (and ensuring that the order keys are added in the
desired order) would correct the ordering problem, but could have
performance implications.

To address this problem, this patch implements a custom serde
serializer for `data::Aggregate` and changes JsonPrinter to use it.
Because serde Serializers are not allowed to directly nest objects, a
new helper struct `WrappedAggregateRow` is introduced to provide a
type that can both have a custom serializer attached as well as pass
a borrowed reference to the columns.